Consolidates token generation and seed scripts to shared package
Moves token generation and database seeding scripts from the database package to a shared scripts package to streamline monorepo organization. Updates documentation and commands accordingly.

Refactors token generation logic for improved maintainability and consistency. Removes redundant files and updates package.json scripts to reflect changes.

Fixes #123 for monorepo script consolidation.
