+++ packages/scripts/package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "build": "tsc --project tsconfig.build.json",
     "dev": "tsc --project tsconfig.build.json --watch",
-    "gen-token": "dotenv -e .env -- tsx src/generate-invitation-token.ts"
+    "gen-token": "dotenv -e .env -- tsx src/generate-invitation-token.ts",
+    "db-seed": "dotenv -e .env -- tsx src/seed.ts"
   },
   "dependencies": {
     "@repo/db": "workspace:*",
