+++ packages/scripts/tsconfig.json
@@ -7,7 +7,7 @@
     "moduleResolution": "bundler",
     "allowImportingTsExtensions": false,
     "noEmit": false
-    // "declaration": false
+
   },
   "include": ["src/**/*"],
   "exclude": ["node_modules", "dist"]
